+++ application/Language/id/Aauth.php
@@ -44,7 +44,7 @@ return [
    'notVerified'            => 'Akun anda belum diverifikasi. Silakan cek email anda dan verifikasi akun anda.',
 
    'loginFailedEmail'       => 'Email dan sandi yang anda masukkan tidak cocok.',
-   'loginFailedName'        => 'Username dan sandi yang Anda masukkan tidak cocok.',
+   'loginFailedUsername'        => 'Username dan sandi yang Anda masukkan tidak cocok.',
    'loginFailedAll'         => 'Email, username dan sandi yang Anda masukkan tidak cocok.',
    'loginAttemptsExceeded'  => 'Anda telah melebihi upaya login anda, akun anda telah diblokir.',