+++ application/Language/sv/Aauth.php
@@ -44,7 +44,7 @@ return [
    'notVerified'            => 'Ditt konto är inte bekräftat. Var vänlig kolla din e-post och bekräfta ditt konto.',
 
    'loginFailedEmail'       => 'E-postadressen och lösenordet stämmer inte överens.',
-   'loginFailedName'        => 'Användarnamnet och lösenordet stämmer inte överens.',
+   'loginFailedUsername'        => 'Användarnamnet och lösenordet stämmer inte överens.',
    'loginFailedAll'         => 'E-postadress, användarnamn och lösenord stämmer inte överens.',
    'loginAttemptsExceeded'  => 'Du har förbrukat dina försök att logga in, ditt konto har blivit låst.',